# docs/openapi.yaml
components:
  schemas:
    AddUserRoleBody:
      description: add user request body
      properties:
        role_id:
          title: Role Id
          type: string
        user_id:
          title: User Id
          type: string
      required:
      - role_id
      - user_id
      title: AddUserRoleBody
      type: object
    Alg:
      description: '1. Enum defining allowed algorithms

        2. some helper functions to make the algorithms accessible'
      enum:
      - ES256
      - RS256
      title: Alg
      type: string
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UserAddBody:
      description: add user request body
      properties:
        alg:
          allOf:
          - $ref: '#/components/schemas/Alg'
          default: ES256
        dob:
          format: date
          title: Dob
          type: string
        email:
          title: Email
          type: string
        first_name:
          title: First Name
          type: string
        password:
          title: Password
          type: string
        postcode:
          title: Postcode
          type: string
        roles:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Roles
        surname:
          title: Surname
          type: string
      required:
      - email
      - password
      - dob
      - postcode
      - first_name
      - surname
      title: UserAddBody
      type: object
    UserUpdateBody:
      description: Update user request body
      properties:
        dob:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Dob
        email:
          title: Email
          type: string
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        new_email:
          anyOf:
          - type: string
          - type: 'null'
          title: New Email
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
        postcode:
          anyOf:
          - type: string
          - type: 'null'
          title: Postcode
        roles:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Roles
        surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Surname
      required:
      - email
      title: UserUpdateBody
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: Jwks Authentication API
  title: Auth Api
  version: '1.0'
openapi: 3.1.0
paths:
...
  /users:
    get:
      description: get all users
      operationId: get_users_users_get
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Users
      tags:
      - Users Authenticated
  /users/me:
    get:
      description: Get user information
      operationId: get_me_users_me_get
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Me
      tags:
      - Users Authenticated
  /users/user:
    delete:
      description: delete a user
      operationId: delete_user_users_user_delete
      parameters:
      - in: query
        name: user_email
        required: true
        schema:
          title: User Email
          type: string
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete User
      tags:
      - Users Authenticated
    patch:
      description: update user endpoint
      operationId: update_user_users_user_patch
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update User
      tags:
      - Users Authenticated
    post:
      description: create new user endpoint. stores a hash of the password
      operationId: create_user_users_user_post
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAddBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create User
      tags:
      - Users Authenticated
  /users/user/roles:
    post:
      description: create new user endpoint. stores a hash of the password
      operationId: add_user_role_users_user_roles_post
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserRoleBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add User Role
      tags:
      - Users Authenticated
...